Qg,acc_flow支持等
**现象及复现步骤**

无

**原因和解决方案**

**影响范围**

无

**验证方案**

无

**涉及分支**

* master

**检查项**

- [ ] 需要同步修改(trident-dump、trident-copy或trident-testsrv)
- [ ] 需要更新依赖
- [ ] 是共性问题(代码中存在类似问题)
- [ ] 打包成功(如果有修改Makefile)

/assign @jinjie

/cc @zhichong @xiangwang

+++ src/common/decapsulate.rs
@@ -10,13 +10,13 @@ use crate::proto::trident::DecapType;
 use crate::utils::bytes;
 
 #[derive(Debug, Clone, Copy, PartialEq, PartialOrd, TryFromPrimitive)]
-#[repr(u16)]
+#[repr(u8)]
 pub enum TunnelType {
-    None = DecapType::None as u16,
-    Vxlan = DecapType::Vxlan as u16,
-    Ipip = DecapType::Ipip as u16,
-    TencentGre = DecapType::Tencent as u16,
-    ErspanOrTeb = TunnelType::TencentGre as u16 + 1,
+    None = DecapType::None as u8,
+    Vxlan = DecapType::Vxlan as u8,
+    Ipip = DecapType::Ipip as u8,
+    TencentGre = DecapType::Tencent as u8,
+    ErspanOrTeb = TunnelType::TencentGre as u8 + 1,
 }
 
 impl From<DecapType> for TunnelType {
